Map Builder Pattern is popping off. Todo is add floor types to the builder, and one more prefab
