+++ src/map/helpers/zone_generation/hub.rs
@@ -1,8 +1,18 @@
-use bevy::{math::Rect, transform::components::Transform};
+use std::collections::HashMap;
 
-use crate::map::components::{EnvironmentalMapCollider, EnvironmentalType, TileType};
+use bevy::{
+    math::{Rect, Vec2},
+    transform::components::Transform,
+};
+use bevy_ecs_tilemap::map::TilemapSize;
 
-use super::map_data::MapData;
+use crate::map::components::{EnvironmentalMapCollider, EnvironmentalType, MarkerType, TileType};
+
+use super::{map_data::MapData, utils::calculate_center_rect};
+
+const PLAYER_SPAWN_Y_OFFSET: f32 = 5.0;
+const LEVEL_EXIT_Y_OFFSET: f32 = 23.0;
+const NPC_OFFSET: f32 = 5.0;
 
 pub fn build_hub(mut map_data: &mut MapData, hub_bounds: &Rect) {
     add_hub_cobblestone(&mut map_data, &hub_bounds);
@@ -206,4 +216,33 @@ fn add_hub_entrance(map_data: &mut MapData, bounds: &Rect) {
             }
         }
     }
+}
+
+pub fn get_hub_markers(
+    map_size: TilemapSize,
+    hub_size: TilemapSize,
+) -> HashMap<MarkerType, Vec<Vec2>> {
+    let mut markers: HashMap<MarkerType, Vec<Vec2>> = HashMap::new();
+
+    let hub_bounds = calculate_center_rect(map_size, hub_size);
+
+    let center_of_hub = hub_bounds.center();
+
+    // Generate player spawn
+    let player_spawn = Vec2::new(center_of_hub.x, hub_bounds.min.y + PLAYER_SPAWN_Y_OFFSET);
+    markers.insert(MarkerType::PlayerSpawns, vec![player_spawn]);
+
+    // Generate level exit
+    let level_exit = Vec2::new(center_of_hub.x, hub_bounds.min.y + LEVEL_EXIT_Y_OFFSET);
+    markers.insert(MarkerType::LevelExits, vec![level_exit]);
+
+    // Generate NPC positions
+    let npc_positions = vec![
+        Vec2::new(center_of_hub.x + NPC_OFFSET, center_of_hub.y + NPC_OFFSET),
+        Vec2::new(center_of_hub.x - NPC_OFFSET, center_of_hub.y - NPC_OFFSET),
+        Vec2::new(center_of_hub.x + NPC_OFFSET, center_of_hub.y - NPC_OFFSET),
+    ];
+    markers.insert(MarkerType::NPCSpawns, npc_positions);
+
+    markers
 }
